test: merge test file && remove judge len(ep)
