+++ zookeeper/resolver.go
@@ -65,9 +65,7 @@ func (z *zookeeperResolver) Resolve(ctx context.Context, desc string) (discovery
 	if err != nil {
 		return discovery.Result{}, err
 	}
-	if len(eps) == 0 {
-		return discovery.Result{CacheKey: desc}, nil
-	}
+
 	instances, err := z.getInstances(eps, path)
 	if err != nil {
 		return discovery.Result{}, err
